redesigning the weapon system and the powerup system
This is kind of a big change all at once but I am rebuilding the powerups to work with the wave end upgrade screen.
I am also working to redesign the weapon system so I can have them drop from asteroids. We are using a base weapon template and resources for the new system.